Remove more closures
